+++ app/src/main/java/com/ecotek/greenshare/FireStore.kt
@@ -1,9 +1,14 @@
 package com.ecotek.greenshare
 
+import android.content.Context
 import android.util.Log
 import android.widget.Toast
+import com.google.firebase.auth.FirebaseAuth
 import com.google.firebase.firestore.FirebaseFirestore
 import com.google.firebase.firestore.SetOptions
+import org.json.JSONArray
+import org.json.JSONException
+import java.io.IOException
 
 class FireStore {
 
@@ -15,10 +20,31 @@ class FireStore {
             .set(userInfo, SetOptions.merge())
         //.addOnSuccessListener { }
     }
+    fun getJsonDataFromAsset(context: Context, fileName: String): List<User>? {
+        val jsonString: String
+        try {
+            jsonString = context.assets.open(fileName).bufferedReader().use { it.readText() }
+            val jsonArray = JSONArray(jsonString)
+            val userList = mutableListOf<User>()
 
-    fun getdata() :String{
-        var currentUsername = ""
-        return currentUsername
+            for (i in 0 until jsonArray.length()) {
+                val jsonObject = jsonArray.getJSONObject(i)
+                val email = jsonObject.getString("email")
+                val password = jsonObject.getString("password")
+                FirebaseAuth.getInstance().createUserWithEmailAndPassword(email,password)
+                FirebaseAuth.getInstance().signOut()
+                val user = User(email, password)
+                userList.add(user)
+            }
+
+            return userList
+        } catch (ioException: IOException) {
+            ioException.printStackTrace()
+            return null
+        } catch (jsonException: JSONException) {
+            jsonException.printStackTrace()
+            return null
+        }
     }
 
     fun getname(activity: FirstFragment,identification : String){
