+++ app/src/main/java/com/ecotek/greenshare/RegisterActivity.kt
@@ -22,35 +22,4 @@ open class RegisterActivity : LoginActivity() {
 
 
     }
-
-    data class User(val email: String, val password: String)
-
-    fun getJsonDataFromAsset(context: Context, fileName: String): List<User>? {
-        val jsonString: String
-        try {
-            jsonString = context.assets.open(fileName).bufferedReader().use { it.readText() }
-            val jsonArray = JSONArray(jsonString)
-            val userList = mutableListOf<User>()
-
-            for (i in 0 until jsonArray.length()) {
-                val jsonObject = jsonArray.getJSONObject(i)
-                val email = jsonObject.getString("email")
-                val password = jsonObject.getString("password")
-                FirebaseAuth.getInstance().createUserWithEmailAndPassword(email,password)
-                FirebaseAuth.getInstance().signOut()
-                val user = User(email, password)
-                userList.add(user)
-            }
-
-            return userList
-        } catch (ioException: IOException) {
-            ioException.printStackTrace()
-            return null
-        } catch (jsonException: JSONException) {
-            jsonException.printStackTrace()
-            return null
-        }
-    }
-
-
 }
